+++ src/infrastructure/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class InfrastructureConfig(AppConfig):
+    name = "src.infrastructure"
+    label = "infrastructure"